backend/requirements.txt
# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Caching & Session Management
redis==5.0.1
aioredis==2.0.1

# Background Tasks
celery==5.3.4
flower==2.0.1

# HTTP Client for OAuth
httpx==0.25.2
authlib==1.2.1

# WebAuthn (Passkeys)
webauthn==1.11.1

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# Template rendering
jinja2==3.1.2

# Email Services
aiosmtplib==3.0.1
sendgrid==6.11.0
aioboto3==12.1.0

# Two-Factor Authentication
pyotp==2.9.0
qrcode[pil]==7.4.2
cryptography==41.0.7